1.0.2 Updates
Updates made for the 1.0.2 release.
Updated the code formatting.
Fixed issue with type method not return correct response, move the
classTypeMap object and for loop to higher level to avoid the overhead
of running the code for each call of .type()
Update the Docs examples
Updated the return for method arrayFirst()
Fixed issue with return of arrayLast() : Now returns as string